+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Advantages:
 - [workflows]: Added npm registry URL and caching to Node.js setup in workflow files.
 - [workflows]: Added concurrency settings in build.yml to cancel previous runs in GitHub Actions.
 - [preset]: Added Thermostat with preset feature. Thanks Ludovic BOUÉ (https://github.com/Luligu/matterbridge/pull/482).
+- [matter.js]: Bump to matter.j v. 0.16.6.
+- [mb_mdns]: Added --ip-filter params to filter incoming mDns messages by sender ip.
 
 ### Changed